b=19312 i=Robert.Read compile warning fix

+++ lustre/tests/mpi/parallel_grouplock.c
@@ -106,15 +106,15 @@ read_buf(int fd)
         rc = read(fd, buf, sizeof(buf));
         if (rc == -1) {
                 pos = lseek(fd, 0, SEEK_CUR);
-                sprintf(errmsg, "read of file %s at pos %d for %lu bytes "
+                sprintf(errmsg, "read of file %s at pos %d for %zu bytes "
                         "returned %d: (%d) %s.\n",
-                        filename, pos, (unsigned long)sizeof(buf), rc, errno, strerror(errno));
+                        filename, pos, sizeof(buf), rc, errno, strerror(errno));
                 FAIL(errmsg);
         } else if (rc != sizeof(buf)) {
                 pos = lseek(fd, 0, SEEK_CUR);
-                sprintf(errmsg, "read of file %s at pos %d for %lu bytes "
+                sprintf(errmsg, "read of file %s at pos %d for %zu bytes "
                         "returned %d.\n",
-                        filename, pos, (unsigned long) sizeof(buf), rc);
+                        filename, pos, sizeof(buf), rc);
                 FAIL(errmsg);
         }
 }
@@ -129,14 +129,14 @@ write_buf(int fd, int index)
         lseek(fd, pos, SEEK_SET);
         rc = write(fd, buf, sizeof(buf));
         if (rc == -1) {
-                sprintf(errmsg, "write of file %s at pos %d for %lu bytes "
+                sprintf(errmsg, "write of file %s at pos %d for %zu bytes "
                         "returned %d: (%d) %s.\n",
-                        filename, pos, (unsigned long)sizeof(buf), rc, errno, strerror(errno));
+                        filename, pos, sizeof(buf), rc, errno, strerror(errno));
                 FAIL(errmsg);
         } else if (rc != sizeof(buf)) {
-                sprintf(errmsg, "write of file %s at pos %d for %lu bytes "
+                sprintf(errmsg, "write of file %s at pos %d for %zu bytes "
                         "returned %d.\n",
-                        filename, pos, (unsigned long)sizeof(buf), rc);
+                        filename, pos, sizeof(buf), rc);
                 FAIL(errmsg);
         }
 }
